move test files in directory `tests`
