--- ruler_test_with_boa3.py
@@ -1,6 +1,0 @@
-from boa3_test.tests.test_classes.testengine import TestEngine
-
-nef_path = 'ruler.nef'
-folder = '.'
-engine = TestEngine(folder)
-engine.run(nef_path, 'addPair')
